+++ R/rwhatsapp.R
@@ -123,7 +123,9 @@ rwa_read <- function(x,
     source = source
   )
 
-  tbl <- rwa_add_emoji(tbl)
+  class(tbl) <- c("rwa_tbl", class(tbl))
+
+  tbl <- lookup_emoji(tbl)
 
   if (verbose) {
     status("emoji extracted")
@@ -306,46 +308,6 @@ rwa_parse_time <- function(time, format, tz) {
 
   return(time)
 }
-
-
-#' @noRd
-#' @importFrom tibble tibble add_column
-#' @importFrom stringi stri_replace_all_regex stri_replace_all_charclass
-#'   stri_split_boundaries
-rwa_add_emoji <- function(x) {
-
-  id <- seq_along(x[["text"]])
-  x <- add_column(x, id = id)
-  text <- x[["text"]]
-
-  text <- stri_replace_all_charclass(text, "[[:punct:][:whitespace:]]", "")
-  l <- stri_split_boundaries(text, type = "character")
-
-  out <- tibble(id = rep(id, sapply(l, length)), emoji = unlist(l))
-
-  out <- add_column(out,
-                    emoji_name = rwhatsapp::emojis$name[
-                      match(out$emoji,
-                            rwhatsapp::emojis$emoji)
-                      ])
-
-  out <- out[!is.na(out$emoji_name), ]
-
-  out <- tibble(id = unique(out$id),
-                emoji = unname(split(out$emoji, out$id)),
-                emoji_name = unname(split(out$emoji_name, out$id)))
-
-  x <- add_column(
-    x,
-    emoji = out$emoji[match(x$id, out$id)],
-    emoji_name = out$emoji_name[match(x$id, out$id)]
-  )
-
-  x$id <- NULL
-
-  return(x)
-}
-
 
 # creates status message and exports start_time if not in parent environment yet
 status <- function(...,